# kyleManagement/contract/kylemanagement.openapi.yaml
...
openapi: 3.0.3
info:
  title: Kyle management API
  description: Internal API
  contact:
    name: Eddy Amegandji
    email: eddysamegandji@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: local development
...
paths:
...
  /tickets:
    get:
      summary: "Get list of tickets"
      tags: ["ticket-resource"]
      operationId: getTickets
      responses:
        '200':
          description: "List retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketDto'
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    post:
      summary: "Create a ticket"
      tags: ["ticket-resource"]
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketDto'
      responses:
        '201':
          description: "A detailed ticket"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDto'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  responses:
    '204':
      description: "No content"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    '401':
      description: "Unauthorized"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    '403':
      description: "Forbidden"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    '404':
      description: "Not found"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        path:
          type: string
        timestamp:
          type: number
          format: double
        message:
          type: string
...
    TicketDto:
      type: object
      properties:
        ticketId:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        creationSummary:
          type: string
        closingDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - TODO
            - RESERVED
            - INPROGRESS
            - PENDING
            - TOBEVALIDATED
            - COMPLETED
            - CANCELED
            - START
        typology:
          type: integer
          format: int64
        initialSkill:
          type: integer
          format: int64
        customer:
          type: integer
          format: int64
        lastModifiedDate:
          type: string
          format: date-time
        lastTicketActionComment:
          type: string
        priority:
          type: string
          enum:
            - VERYLOW
            - LOW
            - MEDIUM
            - HIGH
            - VERYHIGH
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...